Initial commit for branch sync64 of kForth-32. The purpose of this branch is to merge changes to the Forth system code, made in the kForth-64 project, which are source-code compatible with the kForth-32 project. Compatible source code changes are in the src/:{*.h, *.c, *.cpp} files, while the assembly source files src/:{*.s} are not compatible between the two projects.
This branch should be considered experimental and unstable.
Please use the latest stable release from the link below if you
do not want to try out the new features. After a period of
testing and validation, the changes in the sync64 branch will
be merged into the master branch.

https://github.com/mynenik/kForth-32/releases

The version number is set at 1.7.0 for this commit of sync64.

	modified:   src/ForthCompiler.cpp
	modified:   src/ForthCompiler.h
	modified:   src/ForthVM.cpp
	modified:   src/ForthVM.h
	modified:   src/ForthWords.h
	modified:   src/Makefile
	modified:   src/fbc.h
	modified:   src/kforth.cpp
	modified:   src/vm32-common.s
	modified:   src/vm32-fast.s
	modified:   src/vm32.s
	modified:   src/vmc.c