Version 2.0.0 of kForth-32 (sync64 branch)
This version makes a major change to the meaning of "xt" or
execution token and EXECUTE in kForth. Previously, the "xt"
is an address which points to the start of the compiled op-code
sequence (byte code). EXECUTE previously used the pointer to the
start of the byte code.

The revised "xt" adds one level of indirection to the address of
the byte code. The new "xt" is the address of the Cfa field in
the WordListEntry structure (i.e. the word header information
structure) and not the value of the Cfa. EXECUTE has been revised
to dereference the Cfa field address to obtain the Cfa.

The rationale for this change is that, previously (1.x versions),
there was no reliable way to obtain other word header field values
from the "xt". The added indirection in the "xt" allows for
computation of other word header fields, and thereby retrieving
other field values.

In addition to revising EXECUTE , a new word EXECUTE-BC has been
added. EXECUTE-BC (short for "Execute Byte Code") has the behavior
of the older EXECUTE in versions 1.x, and is needed for consistent
implementation of RECURSE and DOES> .

This is a major structural change to kForth-32. Version 2.x series
is experimental and should be considered unstable for now. The
executables, kforth32 and kforth32-fast, satisfactorily pass all of
the system-test code in forth-src/system-test/ . However, they may
not run all of the example programs in forth-src/ correctly. Please
use the latest stable release of the master branch if you want
to ensure correct operation of the examples in forth-src/.

	modified:   ForthCompiler.h
	modified:   ForthVM.cpp
	modified:   ForthWords.h
	modified:   Makefile
	modified:   fbc.h
	modified:   vm32-common.s
	modified:   vm32-fast.s
	modified:   vm32.s