Modified WordList class to be a child of vector<WordListEntry*> rather than vector<WordListEntry>. This allows the extra level of indirection needed to support the generalized execution token "xt" to point to a fixed field address in the word header (WordListEntry) structure. This structural change to the Forth system is needed to support user access to various fields in the word header, for advanced system queries and debugging.
The change to the WordList class necessitates changes to the
system source code in the files listed below. The revised
code builds and kforth32 passes its system tests in
forth-src/system-test , as well as running without errors
a selected group of test Forth source applications. There
are no known bugs at this time. The executable kforth32-fast
has not been checked out with the test code.

	modified:   src/ForthCompiler.cpp
	modified:   src/ForthCompiler.h
	modified:   src/ForthVM.cpp